TeraManager version 0.9.8 August 23th, 2013
--------------------------------------------------------------------------------------------
- Fix: fixed TeraFly's file dialog issue on MacOS.
- Fix: fixed a bug which could cause the TeraConverter's progress bar not to work.
- Fix: fixed some compiler warnings.
- Dev: updated ImageManager / VolumeConverter modules.
--------------------------------------------------------------------------------------------

+++ released_plugins/v3d_plugins/teramanager/src/control/CSettings.cpp
@@ -96,7 +96,7 @@ void CSettings::writeSettings()
 
     settings.beginWriteArray("volumePathHistory");
     std::list<string>::iterator it = volumePathHistory.begin();
-    for (int i = 0; i < volumePathHistory.size(); ++i, it++) {
+    for (size_t i = 0; i < volumePathHistory.size(); ++i, it++) {
         settings.setArrayIndex(i);
         QString path(it->c_str());
         settings.setValue("path", path);
